sql_migration_assistant/requirements.txt
databricks-sdk
gradio==4.27.0
pyyaml
scipy
mlflow[databricks]
databricks-labs-blueprint